+++ CHANGELOG.md
@@ -7,6 +7,10 @@ The format is based on [Keep a Changelog][changelog] and this project adheres to
 
 ## Unreleased
 
+### Added
+
+- The `clean` command to delete files which are not found in a snapshot
+
 ## 12.0.0 - 2022-01-08
 
 ### Changed